Add merge command
