+++ cli/package.json
@@ -32,6 +32,7 @@
     "chalk": "^5.0.1",
     "clipboardy": "^3.0.0",
     "cosmiconfig": "^7.0.1",
+    "deepmerge": "^4.2.2",
     "dotenv": "^16.0.1",
     "enquirer": "^2.3.6",
     "lodash.get": "^4.4.2",
